refactor: modify packaging for enum folder
